sec_agents/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "sec-agents"
version = "0.1.0"
description = "Cybersecurity AI agents for threat analysis and response"
authors = [{name = "Security Team"}]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "langchain>=0.1.0",
    "langgraph>=0.0.60",
    "openai>=1.3.0",
    "anthropic>=0.7.0",
    "pydantic>=2.0.0",
    "pytest>=7.4.0",
    "pyyaml>=6.0.1",
    "click>=8.1.0",
    "httpx>=0.25.0",
    "pandas>=2.0.0",
    "jupyter>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.5.0",
    "ruff>=0.1.0",
]

[project.scripts]
sec-agents = "sec_agents.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["sec_agents"]

[tool.black]
line-length = 88
target-version = ['py310']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.ruff]
line-length = 88
target-version = "py310"
select = ["E", "F", "W", "B", "I"]